+++ qbraid/transpiler/conversions/cirq/__init__.py
@@ -29,9 +29,10 @@ Functions
    cirq_to_pyqir
    cirq_to_stim
    stim_to_cirq
+   cirq_to_qat
 
 """
-from .cirq_extras import cirq_to_pyqir, cirq_to_stim, stim_to_cirq
+from .cirq_extras import cirq_to_pyqir, cirq_to_qat, cirq_to_stim, stim_to_cirq
 from .cirq_to_braket import cirq_to_braket
 from .cirq_to_pyquil import cirq_to_pyquil
 from .cirq_to_qasm2 import cirq_to_qasm2
@@ -43,4 +44,5 @@ __all__ = [
     "cirq_to_pyqir",
     "cirq_to_stim",
     "stim_to_cirq",
+    "cirq_to_qat",
 ]